fix: restrict access to admin routes and create adminMiddleware
Added adminMiddleware to ensure only users with role 'admin' can access admin-related routes.
