+++ routes/user.js
@@ -1,10 +1,10 @@
-import express from 'express';
-const router = express.Router();
+// import express from 'express';
+// const router = express.Router();
 
-import authenticateToken from '../middleware/authMiddleware.js';
-import { getAllUsers } from '../controllers/userController.js';
+// import authenticateToken from '../middleware/authMiddleware.js';
+// import { getAllUsers } from '../controllers/userController.js';
 
-// Endpoint untuk mendapatkan semua user
-router.get('/users', authenticateToken, getAllUsers);
+// // Endpoint untuk mendapatkan semua user
+// router.get('/users', authenticateToken, getAllUsers);
 
-export default router;
+// export default router;
